Enable queueing in the tree descent (#56)
* Descending the source tree generates tasks on the queue.
* Remove hashing decorater in favour of a hashing task.
* Rename `language` module to `tasks`.
* Move concrete classes out of `tasks/__init__.py`.
